Move client side event listeners into Layout
