+++ pages/contact-us/[id].tsx
@@ -1,4 +1,4 @@
-import { Fragment, useEffect, useCallback, useMemo, useState } from 'react'
+import { Fragment } from 'react'
 import TableContents from '../../components/TableContents'
 import Heading from '../../components/Heading'
 
@@ -20,8 +20,6 @@ import {
 } from '../../lib/auth'
 import { authOptions } from '../api/auth/[...nextauth]'
 import { getServerSession } from 'next-auth/next'
-import throttle from 'lodash.throttle'
-import { useRouter } from 'next/router'
 import { GetServerSideProps } from 'next'
 import { getToken } from 'next-auth/jwt'
 
@@ -47,56 +45,6 @@ interface ContactUsPageProps {
 
 const ContactUsPage = (props: ContactUsPageProps) => {
   /* istanbul ignore next */
-
-  const [response, setResponse] = useState<Response | undefined>()
-  const router = useRouter()
-
-  const validationResponse = useCallback(
-    async () => setResponse(await fetch('/api/refresh-msca')),
-    [],
-  )
-  //Event listener for click events that revalidates MSCA session, throttled using lodash to only trigger every 1 minute
-  const onClickEvent = useCallback(
-    async () => setResponse(await fetch('/api/refresh-msca')),
-    [],
-  )
-  const throttledOnClickEvent = useMemo(() => {
-    return throttle(onClickEvent, 60000, { trailing: false })
-  }, [onClickEvent])
-
-  const throttledOnClickEvent = useMemo(
-    () => throttle(validationResponse, 60000, { trailing: false }),
-    [validationResponse],
-  )
-  //Event listener for visibility change events that revalidates MSCA session, throttled using lodash to only trigger every 15 seconds
-  const throttledVisiblityChangeEvent = useMemo(
-    () => throttle(validationResponse, 15000, { trailing: false }),
-    [validationResponse],
-  )
-
-  //If session is valid, add event listeners to check for valid sessions on click and visibility change events
-  useEffect(() => {
-    window.addEventListener('visibilitychange', throttledVisiblityChangeEvent)
-    window.addEventListener('click', throttledOnClickEvent)
-    //If validateSession call indicates an invalid MSCA session, redirect to logout
-    if (response?.status === 401) {
-      router.push(`/${props.locale}/auth/logout`)
-    }
-    //Remove event on unmount to prevent a memory leak with the cleanup
-    return () => {
-      window.removeEventListener('click', throttledOnClickEvent)
-      window.removeEventListener(
-        'visiblitychange',
-        throttledVisiblityChangeEvent,
-      )
-    }
-  }, [
-    throttledOnClickEvent,
-    throttledVisiblityChangeEvent,
-    response,
-    router,
-    props.locale,
-  ])
 
   return (
     <div
